Remove interceptor and create instance in constructor - COMMENT: Call HTTP put from client - possibly remove and make public in CouchDbClient class? - TODO: Remove uri getUserInfo logic - Remove <T> T executeRequest call in database - Move database from CouchDbClient and remove CouchDbClient class - Rename CouchDbClientBase to CouchDbClient - Call AgentHeader class in CloudantClientTests - Change executeToInputStream method to public in CouchDbClient - Remove custom headers - Use IOUtils.closeQuietly for close method - Remove '/' in URI from Database's shards method
